move sidebar state to sidebar component
