+++ cypress.json
@@ -1,6 +1,5 @@
 {
     "component": {
-        "componentFolder": "src",
         "testFiles": "**/*spec.{js,jsx,ts,tsx}",
         "video": false,
         "screenshotOnRunFailure": false
