--- cypress/integration/sample_spec.tsx
@@ -1,13 +1,0 @@
-describe("My First Test", () => {
-  it("Visits the Kitchen Sink", () => {
-    cy.visit("https://example.cypress.io");
-
-    cy.contains("type").click();
-
-    cy.url().should("include", "commands/actions");
-
-    cy.get(".action-email")
-      .type("fake@email.com")
-      .should("have.value", "fake@email.com");
-  });
-});
