+++ src/layout/SideBarHeader/styles.module.scss
@@ -3,7 +3,7 @@
 .header {
     @include flex-common(space-between, center);
 
-    padding: 0 0.25rem;
+    padding: 0 0.5rem;
 }
 
 .menu {
