+++ src/styles/scss/_variables.scss
@@ -31,3 +31,4 @@ $orange-300: #ffbc74;
 $orange-400: #ffddb9;
 $orange-500: #fff3e7;
 $header-height: 65px;
+$sidebar-width: 250px;
